+++ archive/zipfile/doc.go
@@ -1,0 +1,2 @@
+// Package zipfile contains zip-related helper functions.
+package zipfile