+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 ## Unreleased - YYYY-MM-DD
 - Added io/file package
+- Added archive/zipfile package